refactor: split input.c into multiple files
