+++ Makefile
@@ -71,10 +71,12 @@ libs: \
 	$(OUTPUT_DIR)/libs/libkoreader-cre.so \
 	$(OUTPUT_DIR)/libs/libwrap-mupdf.so
 
-$(OUTPUT_DIR)/libs/libkoreader-input.so: input.c $(POPEN_NOSHELL_LIB)
-	$(CC) $(DYNLIB_CFLAGS) -I$(POPEN_NOSHELL_DIR) \
+$(OUTPUT_DIR)/libs/libkoreader-input.so: input/*.c input/*.h $(POPEN_NOSHELL_LIB)
+	@echo "Building koreader input module..."
+	$(CC) $(DYNLIB_CFLAGS) -I$(POPEN_NOSHELL_DIR) -I./input \
 		$(if $(KOBO),-DKOBO,) $(if $(KINDLE),-DKINDLE,) $(if $(POCKETBOOK),-DPOCKETBOOK,) \
-		-o $@ $^ $(if $(POCKETBOOK),-linkview,)
+		-o $@ \
+		input/input.c $(POPEN_NOSHELL_LIB) $(if $(POCKETBOOK),-linkview,)
 
 $(OUTPUT_DIR)/libs/libkoreader-lfs.so: \
 			$(if $(or $(ANDROID),$(WIN32)),$(LUAJIT_LIB),) \
