+++ input/input.h
@@ -15,8 +15,8 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
-#ifndef _PDF_INPUT_H
-#define _PDF_INPUT_H
+#ifndef _KO_INPUT_H
+#define _KO_INPUT_H
 
 #include <lua.h>
 #include <lualib.h>
